@@ -15,7 +15,7 @@
     "composer/ca-bundle": "^1.2",
     "doctrine/data-fixtures": "^1.3",
     "doctrine/dbal": "^2.8",
-    "doctrine/migrations": "^2",
+    "doctrine/migrations": "^3",
     "doctrine/orm": "~2.6",
     "guzzlehttp/guzzle": ">6.0",
     "http-interop/http-factory-guzzle": "^1.0",
@@ -26,22 +26,29 @@
     "monolog/monolog": "^2",
     "php-di/php-di": "^6.0",
     "php-di/slim-bridge": "^3.0",
-    "slim/http": "^0.8",
+    "slim/http": "^1",
     "slim/slim": "^4.2",
-    "symfony/console": "^4.3",
-    "symfony/event-dispatcher": "^4.3",
-    "symfony/serializer": "^4.3",
-    "symfony/validator": "^4.3"
+    "symfony/console": "^5",
+    "symfony/event-dispatcher": "^5",
+    "symfony/serializer": "^5",
+    "symfony/validator": "^5"
   },
   "require-dev": {
     "roave/security-advisories": "dev-master",
     "codeception/codeception": "^4.0",
     "codeception/module-phpbrowser": "^1.0",
     "mockery/mockery": "^1.0",
-    "phpstan/phpstan": "^0.11.1",
-    "phpstan/phpstan-strict-rules": "^0.11.0",
+    "phpstan/phpstan": "^0.12",
+    "phpstan/phpstan-doctrine": "^0.12",
+    "phpstan/phpstan-strict-rules": "^0.12",
     "overtrue/phplint": "^1.1",
     "codeception/module-asserts": "^1.1"
+  },
+  "replace": {
+    "symfony/polyfill-php71": "1.99",
+    "symfony/polyfill-php72": "1.99",
+    "symfony/polyfill-php73": "1.99",
+    "symfony/polyfill-php74": "1.99"
   },
   "scripts": {
     "test": [
