@@ -17,22 +17,22 @@ use function random_bytes;
 class Assets
 {
     /** @var array Known libraries loaded in initialization. */
-    protected $libraries = [];
+    protected array $libraries = [];
 
     /** @var array An optional array lookup for versioned files. */
-    protected $versioned_files = [];
+    protected array $versioned_files = [];
 
     /** @var array Loaded libraries. */
-    protected $loaded = [];
+    protected array $loaded = [];
 
     /** @var bool Whether the current loaded libraries have been sorted by order. */
-    protected $is_sorted = true;
+    protected bool $is_sorted = true;
 
     /** @var string A randomly generated number-used-once (nonce) for inline CSP. */
     protected $csp_nonce;
 
     /** @var array The loaded domains that should be included in the CSP header. */
-    protected $csp_domains;
+    protected array $csp_domains;
 
     /**
      * Assets constructor.
@@ -42,8 +42,10 @@ class Assets
      *
      * @throws \Exception
      */
-    public function __construct(array $libraries = [], array $versioned_files = [])
-    {
+    public function __construct(
+        array $libraries = [],
+        array $versioned_files = []
+    ) {
         foreach ($libraries as $library_name => $library) {
             $this->addLibrary($library, $library_name);
         }
@@ -202,10 +204,11 @@ class Assets
      * Add a single javascript inline script.
      *
      * @param string|array $js_script
+     * @param int $order
      *
      * @return $this
      */
-    public function addInlineJs($js_script, $order = 100): self
+    public function addInlineJs($js_script, int $order = 100): self
     {
         $this->load([
             'order' => $order,
@@ -221,10 +224,11 @@ class Assets
      * Add a single CSS file.
      *
      * @param string|array $css_script
+     * @param int $order
      *
      * @return $this
      */
-    public function addCss($css_script, $order = 100): self
+    public function addCss($css_script, int $order = 100): self
     {
         $this->load([
             'order' => $order,
@@ -261,7 +265,7 @@ class Assets
      * Returns all CSS includes and inline styles.
      * @return string HTML tags as string.
      */
-    public function css()
+    public function css(): string
     {
         $this->_sort();
 
@@ -294,7 +298,7 @@ class Assets
      * Returns all script include tags.
      * @return string HTML tags as string.
      */
-    public function js()
+    public function js(): string
     {
         $this->_sort();
 
@@ -346,7 +350,7 @@ class Assets
     /**
      * Sort the list of loaded libraries.
      */
-    protected function _sort()
+    protected function _sort(): void
     {
         if (!$this->is_sorted) {
             uasort($this->loaded, function ($a, $b) {
