@@ -7,8 +7,7 @@ use function is_string;
 
 class EventDispatcher extends \Symfony\Component\EventDispatcher\EventDispatcher
 {
-    /** @var CallableResolverInterface */
-    protected $callableResolver;
+    protected CallableResolverInterface $callableResolver;
 
     public function __construct(CallableResolverInterface $callableResolver)
     {
@@ -17,7 +16,7 @@ class EventDispatcher extends \Symfony\Component\EventDispatcher\EventDispatcher
         $this->callableResolver = $callableResolver;
     }
 
-    public function addServiceSubscriber($class_name)
+    public function addServiceSubscriber($class_name): void
     {
         if (is_array($class_name)) {
             foreach ($class_name as $service) {
@@ -31,17 +30,17 @@ class EventDispatcher extends \Symfony\Component\EventDispatcher\EventDispatcher
                 $this->addListener($eventName, $this->_getCallable($class_name, $params));
             } elseif (is_string($params[0])) {
                 $this->addListener($eventName, $this->_getCallable($class_name, $params[0]),
-                    isset($params[1]) ? $params[1] : 0);
+                    $params[1] ?? 0);
             } else {
                 foreach ($params as $listener) {
                     $this->addListener($eventName, $this->_getCallable($class_name, $listener[0]),
-                        isset($listener[1]) ? $listener[1] : 0);
+                        $listener[1] ?? 0);
                 }
             }
         }
     }
 
-    public function removeServiceSubscriber($class_name)
+    public function removeServiceSubscriber($class_name): void
     {
         if (is_array($class_name)) {
             foreach ($class_name as $service) {
@@ -62,7 +61,7 @@ class EventDispatcher extends \Symfony\Component\EventDispatcher\EventDispatcher
         }
     }
 
-    protected function _getCallable($class_name, $method)
+    protected function _getCallable($class_name, $method): DeferredCallable
     {
         return new DeferredCallable($class_name . ':' . $method, $this->callableResolver);
     }
