@@ -1,17 +1,13 @@
 <?php
-use App\Event;
+
 use App\Console\Command;
-use App\Settings;
+use App\Event;
 use App\Middleware;
-use Doctrine\Migrations\Configuration\Configuration;
-use Doctrine\Migrations\Tools\Console\ConsoleRunner;
-use Doctrine\Migrations\Tools\Console\Helper\ConfigurationHelper;
-use Doctrine\ORM\EntityManager;
+use App\Settings;
 use Slim\Interfaces\ErrorHandlerInterface;
 
-return function (\App\EventDispatcher $dispatcher)
-{
-    $dispatcher->addListener(Event\BuildConsoleCommands::class, function(Event\BuildConsoleCommands $event) {
+return function (\App\EventDispatcher $dispatcher) {
+    $dispatcher->addListener(Event\BuildConsoleCommands::class, function (Event\BuildConsoleCommands $event) {
         $console = $event->getConsole();
         $di = $console->getContainer();
 
@@ -28,35 +24,28 @@ return function (\App\EventDispatcher $dispatcher)
             Doctrine\ORM\Tools\Console\ConsoleRunner::addCommands($console);
 
             // Add Doctrine Migrations
-            $defaults = [
-                'table_name' => 'app_migrations',
-                'directory' => $settings[Settings::BASE_DIR] . '/src/Entity/Migration',
-                'namespace' => 'App\Entity\Migration',
-            ];
-
-            $user_options = $settings[Settings::DOCTRINE_OPTIONS]['migrations'] ?? [];
-            $options = array_merge($defaults, $user_options);
-
-            /** @var EntityManager $em */
-            $em = $di->get(EntityManager::class);
-            $connection = $em->getConnection();
+            /** @var Doctrine\ORM\EntityManagerInterface $em */
+            $em = $di->get(Doctrine\ORM\EntityManagerInterface::class);
 
             $helper_set = $console->getHelperSet();
             $doctrine_helpers = Doctrine\ORM\Tools\Console\ConsoleRunner::createHelperSet($em);
-
             $helper_set->set($doctrine_helpers->get('db'), 'db');
             $helper_set->set($doctrine_helpers->get('em'), 'em');
 
-            $migrate_config = new Configuration($connection);
-            $migrate_config->setMigrationsTableName($options['table_name']);
-            $migrate_config->setMigrationsDirectory($options['directory']);
-            $migrate_config->setMigrationsNamespace($options['namespace']);
+            $migrateConfig = new Doctrine\Migrations\Configuration\Migration\ConfigurationArray([
+                'migrations_paths' => [
+                    'App\Entity\Migration' => $settings[Settings::BASE_DIR] . '/src/Entity/Migration',
+                ],
+                'table_storage' => [
+                    'table_name' => 'app_migrations',
+                ],
+            ]);
 
-            $migrate_config_helper = new ConfigurationHelper($connection,
-                $migrate_config);
-            $helper_set->set($migrate_config_helper, 'configuration');
-
-            ConsoleRunner::addCommands($console);
+            $migrateFactory = Doctrine\Migrations\DependencyFactory::fromEntityManager(
+                $migrateConfig,
+                new Doctrine\Migrations\Configuration\EntityManager\ExistingEntityManager($em)
+            );
+            Doctrine\Migrations\Tools\Console\ConsoleRunner::addCommands($console, $migrateFactory);
         }
 
         if (file_exists(__DIR__ . '/cli.php')) {
@@ -64,7 +53,7 @@ return function (\App\EventDispatcher $dispatcher)
         }
     });
 
-    $dispatcher->addListener(Event\BuildRoutes::class, function(Event\BuildRoutes $event) {
+    $dispatcher->addListener(Event\BuildRoutes::class, function (Event\BuildRoutes $event) {
         $app = $event->getApp();
 
         // Load app-specific route configuration.
