@@ -3,7 +3,7 @@ namespace App\Http;
 
 use Psr\Http\Message\ResponseInterface;
 
-class Response extends \Slim\Http\Response
+final class Response extends \Slim\Http\Response
 {
     public const CACHE_ONE_MINUTE = 60;
     public const CACHE_ONE_HOUR = 3600;
@@ -69,7 +69,7 @@ class Response extends \Slim\Http\Response
      */
     public function withJson($data, ?int $status = null, int $options = 0, int $depth = 512): ResponseInterface
     {
-        $options &= JSON_UNESCAPED_SLASHES;
+        $options |= JSON_UNESCAPED_SLASHES;
 
         return parent::withJson($data, $status, $options, $depth);
     }
